Remove unneeded include, fix warning
2008-05-08  Bastien Nocera  <hadess@hadess.net>

	* plparse/totem-pl-parser-pla.c (totem_pl_parser_add_pla):
	Remove unneeded include, fix warning

	* plparse/totem-pl-parser-private.h:
	* plparse/totem-pl-parser.c (totem_pl_parser_relative),
	(totem_pl_parser_ignore_from_mimetype):
	Update totem_pl_parser_relative() to take in a GFile as the output
	target, and simplify
	Remove old code in totem_pl_parser_ignore_from_mimetype()

	* plparse/test-parser.c (test_relative_real):
	Fix usage of totem_pl_parser_relative()

	* plparse/totem-pl-parser-lines.c (totem_pl_parser_url_to_dos),
	(totem_pl_parser_write_m3u):
	* plparse/totem-pl-parser-pls.c (totem_pl_parser_write_pls):
	* plparse/totem-pl-parser-xspf.c (totem_pl_parser_write_xspf):
	Use totem_pl_parser_relative() where needed, and fix usage


svn path=/trunk/; revision=117

+++ plparse/totem-pl-parser-pla.c
@@ -26,7 +26,6 @@
 #include <glib.h>
 #include <glib/gi18n-lib.h>
 #include <gtk/gtk.h>
-#include <libgnomevfs/gnome-vfs.h>
 #include "totem-pl-parser.h"
 #include "totemplparser-marshal.h"
 #endif /* !TOTEM_PL_PARSER_MINI */
@@ -160,7 +159,7 @@ totem_pl_parser_add_pla (TotemPlParser *parser,
 	if (size < RECORD_SIZE)
 	{
 		g_free (contents);
-		DEBUG(file, g_print ("playlist '%s' is too short: %d\n", uri, size));
+		DEBUG(file, g_print ("playlist '%s' is too short: %d\n", uri, (unsigned int) size));
 		return TOTEM_PL_PARSER_RESULT_ERROR;
 	}
